+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 Used [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.18.0] - 2021-10-08
+### Added
+- Methods beneficiaryList, beneficiaryAdd, beneficiaryEdit, beneficiaryBlock, beneficiaryUnblock, beneficiaryShow
+- Parameter $beneficiaryId in methods cardRefill, paymentToRegisteredCard, paymentToAccount, paymentToAccountWithTaxData, paymentToUnregisteredCard, paymentToUnregisteredCardWithForm
+
 ## [v1.17.0] - 2021-10-05
 ### Added
 - Method bindSelfemployments